修复射频前端控制裸机

+++ pcie/rf/rf_ctrl.h
@@ -28,8 +28,8 @@
 #define C_ATTEN_STEP 0.25
 #define C_ATTEN_MIN_VAL 10.0f
 #define C_STAGE_ATTEN_SET_MAX_VAL 0xff
-#define C_ATTEN_SET_MAX_VAL (uint16_t)(C_STAGE_ATTEN_SET_MAX_VAL * 2)
-#define C_ATTEN_SET_MIN_VAL (uint16_t)(C_ATTEN_MIN_VAL / C_ATTEN_STEP)
+#define C_ATTEN_SET_MAX_VAL (uint8_t)(C_STAGE_ATTEN_SET_MAX_VAL * 2)
+#define C_ATTEN_SET_MIN_VAL (uint8_t)(C_ATTEN_MIN_VAL / C_ATTEN_STEP)
 
 // 生成指定位宽的掩码
 #define C_FIELD_MASK_RAW(bitWidth) ((1 << bitWidth) - 1)
